requirements.txt
pyTelegramBotAPI==4.16.1
pydantic==1.10.14
loguru==0.7.2
sqlmodel==0.0.8
psycopg2-binary==2.9.7
python-telegram-bot-calendar==1.0.5

